fix: align handler field compliance ownership
